requirements.txt
flask
groq
serverless-wsgi
python-dotenv
google-genai
pydantic